refactor(core): remove proxy layer and enforce direct module imports
