+++ training/train_lstm.py
@@ -9,7 +9,7 @@ import datetime
 import joblib
 import numpy as np
 from sklearn.preprocessing import MinMaxScaler
-from core.services import StockPriceFetcher
+from core.data.data_fetcher import StockPriceFetcher
 from app.models.lstm_model import build_lstm_model
 
 LOOKBACK_WINDOW = 60
